implement userRoute userModel userController and authMiddleware

+++ backend/controllers/goalController.js
@@ -1,19 +1,14 @@
 import asyncHandler from 'express-async-handler';
 import Goal from '../models/goalModel.js';
+import User from '../models/userModel.js';
 
-// @desc    Get all goals
-// @route   GET /api/goals
-// @access  Private
-const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+export const getGoals = asyncHandler(async (req, res) => {
+    const goals = await Goal.find({ user: req.user.id });
 
     res.status(200).json(goals);
 });
 
-// @desc    Create a goal
-// @route   POST /api/goals
-// @access  Private
-const setGoal = asyncHandler(async (req, res) => {
+export const setGoal = asyncHandler(async (req, res) => {
     if (!req.body.text) {
         res.status(400);
         throw new Error('please add a goal');
@@ -21,20 +16,30 @@ const setGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     });
 
     res.status(201).json(goal);
 });
 
-// @desc    Update a goal
-// @route   PUT /api/goals/:id
-// @access  Private
-const updateGoal = asyncHandler(async (req, res) => {
+export const updateGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id);
 
     if (!goal) {
         res.status(404);
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -44,10 +49,7 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(200).json(updatedGoal);
 });
 
-// @desc    Delete a goal
-// @route   DELETE /api/goals/:id
-// @access  Private
-const deleteGoal = asyncHandler(async (req, res) => {
+export const deleteGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id);
 
     if (!goal) {
@@ -55,9 +57,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found');
     }
 
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
+    }
+
     await Goal.findByIdAndDelete(req.params.id)
 
     res.status(200).json({ id: req.params.id });
 });
-
-export { getGoals, setGoal, updateGoal, deleteGoal };
